reload_info constructor
